+++ aiodocker/networks.py
@@ -1,12 +1,13 @@
 import json
 from .utils import clean_filters
+from typing import Any, Dict, List, Mapping
 
 
 class DockerNetworks:
     def __init__(self, docker):
         self.docker = docker
 
-    async def list(self, *, filters=None):
+    async def list(self, *, filters: Mapping = None) -> List[Dict[str, Any]]:
         """
         Return a list of networks
 
@@ -27,14 +28,14 @@ class DockerNetworks:
         data = await self.docker._query_json("networks", params=params)
         return data
 
-    async def create(self, config):
-        config = json.dumps(config, sort_keys=True).encode("utf-8")
+    async def create(self, config: Dict[str, Any]) -> "DockerNetwork":
+        bconfig = json.dumps(config, sort_keys=True).encode("utf-8")
         data = await self.docker._query_json(
-            "networks/create", method="POST", data=config
+            "networks/create", method="POST", data=bconfig
         )
         return DockerNetwork(self.docker, data["Id"])
 
-    async def get(self, net_specs):
+    async def get(self, net_specs: str) -> "DockerNetwork":
         data = await self.docker._query_json(
             "networks/{net_specs}".format(net_specs=net_specs), method="GET"
         )
@@ -46,26 +47,26 @@ class DockerNetwork:
         self.docker = docker
         self.id = id_
 
-    async def show(self):
+    async def show(self) -> Dict[str, Any]:
         data = await self.docker._query_json("networks/{self.id}".format(self=self))
         return data
 
-    async def delete(self):
+    async def delete(self) -> bool:
         async with self.docker._query(
             "networks/{self.id}".format(self=self), method="DELETE"
-        ):
-            pass
+        ) as resp:
+            return resp.status == 204
 
-    async def connect(self, config):
-        config = json.dumps(config, sort_keys=True).encode("utf-8")
+    async def connect(self, config: Dict[str, Any]) -> None:
+        bconfig = json.dumps(config, sort_keys=True).encode("utf-8")
         await self.docker._query_json(
-            "networks/{self.id}/connect".format(self=self), method="POST", data=config
+            "networks/{self.id}/connect".format(self=self), method="POST", data=bconfig
         )
 
-    async def disconnect(self, config):
-        config = json.dumps(config, sort_keys=True).encode("utf-8")
+    async def disconnect(self, config: Dict[str, Any]) -> None:
+        bconfig = json.dumps(config, sort_keys=True).encode("utf-8")
         await self.docker._query_json(
             "networks/{self.id}/disconnect".format(self=self),
             method="POST",
-            data=config,
+            data=bconfig,
         )